refactor: build-tag sparse detection; docs: fix markdownlint MD038 (CodeRabbit/Copilot)
- Move sparseInfo behind build tags (sparse_unix.go for unix, sparse_other.go
  no-op for !unix) so internal/engine no longer type-asserts syscall.Stat_t in
  a platform-agnostic file — flagged by three reviewers. Verified the engine
  package now cross-compiles for GOOS=windows (a remaining windows failure is
  pre-existing in internal/storage, out of scope). Linux/macOS behaviour
  unchanged.
- CONTRIBUTING.md: drop trailing spaces inside `### ` / `- ` code spans (MD038).
